Refactor `get_post_preview_id` to Utils
