+++ src/Data/NodeResolver.php
@@ -7,6 +7,7 @@ use WP_Post;
 use WPGraphQL\AppContext;
 use GraphQL\Error\UserError;
 use WPGraphQL\Router;
+use WPGraphQL\Utils\Utils;
 
 class NodeResolver {
 
@@ -235,7 +236,7 @@ class NodeResolver {
 			$post_id = $queried_object->ID;
 
 			if ( isset( $extra_query_vars['asPreview'] ) && true === $extra_query_vars['asPreview'] ) {
-				$post_id = self::get_post_preview_id($post_id);
+				$post_id = Utils::get_post_preview_id($post_id);
 			}
 
 			return ! empty( $post_id ) ? $this->context->get_loader( 'post' )->load_deferred( $post_id ) : null;
@@ -586,30 +587,6 @@ class NodeResolver {
 		do_action_ref_array( 'parse_request', [ &$this->wp ] );
 
 		return $uri;
-	}
-
-	/**
-	 * Given a WP Post or post ID, this method attempts to resolve a preview post ID.
-	 * 
-	 * @param Int|WP_Post $post The WP Post object or Post ID
-	 * 
-	 * @return Int A preview post ID if one exists, the current post ID if one doesn't exist.
-	 */
-	public static function get_post_preview_id( $post ) {
-		$post_id = is_object($post) ? $post->ID : $post;
-
-		$revisions = wp_get_post_revisions(
-			$post_id,
-			[
-				'posts_per_page' => 1,
-				'fields'         => 'ids',
-				'check_enabled'  => false,
-			]
-		);
-
-		$post_id = ! empty( $revisions ) ? array_values( $revisions )[0] : $post_id;
-
-		return $post_id;
 	}
 
 	/**
